requirements/nn.txt
filelock==3.24.3
fsspec==2026.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
pillow==12.1.1
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
